work on edit game dialog

+++ src/components/adminTools/edit/EditGameDialog.js
@@ -1,14 +1,16 @@
 import { Button, Dialog, DialogActions, DialogContent, DialogContentText, DialogTitle, Grid, makeStyles, Paper, TextField, Typography } from "@material-ui/core";
+import { Alert } from "@mui/material";
 import DeleteIcon from '@mui/icons-material/Delete';
 //form components
 import DatePickerDesktop from '../DatePickerDesktop';
 import TeamSearch from "../TeamSearch";
-import HomeTeamRadio from "../HomeTeamRadio";
+// import HomeTeamRadio from "../HomeTeamRadio";
 import SeasonSelect from "../../seasonProfile/SeasonSelect";
 import { useState } from "react";
 //APIs
 import { deleteFight as deleteFightAPI } from "../../../api/fights/fightApi";
-import { Alert } from "@mui/material";
+import { updateGame as updateGameAPI } from "../../../api/games/gamesApi";
+
 
 const useStyles = makeStyles((theme) => ({
     gameTypeTextField: {
@@ -33,6 +35,9 @@ const useStyles = makeStyles((theme) => ({
     fightPaper: {
         padding: '10px',
         display: 'flex'
+    },
+    deleteIcon: {
+        cursor: 'pointer'
     }
 }));
 
@@ -52,41 +57,81 @@ const EditGameDialog = ({ fights, setFights, game, open, handleClose }) => {
     const [deleteSuccessful, setDeleteSuccessful] = useState(false);
     const [deleteUnsuccessful, setDeleteUnsuccessful] = useState(false);
 
+    //succesful updategame
+    const[updateSuccessful, setUpdateSuccessful] = useState(false);
+    const[updateUnsuccessful, setUpdateUnsuccessful] = useState(false);
+
     const updateGame = () => {
-        let gameInfo = {
-            date: dateOfGame,
-            season,
-            teams: gameTeams,
-            gameType: typeOfGame,
-            description: gameDescription
+        let gameInfo = {};
+
+        if(season !== game.season.season){
+            gameInfo.season = season;
         }
+        if(dateOfGame !== game.date){
+            gameInfo.date = dateOfGame;
+        }
+        if(gameTeams !== game.teams){
+            gameInfo.teams = gameTeams;
+        }
+        if(typeOfGame !== game.gameType){
+            gameInfo.gameType = typeOfGame;
+        }
+        if(gameDescription !== game.description) {
+            gameInfo.description = gameDescription;
+        }
+
         console.log(gameInfo);
+
+        updateGameAPI(game._id, gameInfo).then(response => {
+            console.log(response);
+            if(response.success === true){
+                setUpdateSuccessful(true);
+                setSeason(response.data.season);
+                setDateOfgame(response.data.date);
+                setGameTeams(response.data.teams);
+                setTypeOfGame(response.data.gameType);
+                setGameDescription(response.data.description);
+
+                setTimeout(() => {
+                    setUpdateSuccessful(false);
+                    handleClose();
+                }, 2000);
+            } else {
+                setUpdateUnsuccessful(true);
+                setTimeout(() => {
+                    setUpdateUnsuccessful(false);
+                    handleClose();
+                }, 2000);
+            }
+        });        
     }
 
     const updateTeams = (newTeam, oldTeam) => {
+        console.log(newTeam, oldTeam)
         let teams = [...gameTeams];
-        let newTeams = teams.filter(team => {
-            return team.id !== oldTeam.id;
-        });
+        let newTeams = teams.filter(team => 
+            team.id !== oldTeam.id
+        );
         newTeam.id = newTeam._id;
         delete newTeam._id;
         newTeams.push(newTeam);
+        console.log(newTeams);
 
         setGameTeams(newTeams);
     }
 
-    const updateHomeTeam = (teamId) => {
-        // console.log(teamId);
-        let teams = [...gameTeams];
-        if(teams[0].id === teamId){
-            teams[0].home = true;
-            teams[1].home = false;
-        } else {
-            teams[1].home = true;
-            teams[0].home = false;
-        }
-        setGameTeams(teams);
-    }
+    // const updateHomeTeam = (teamId) => {
+    //     // console.log(teamId);
+    //     let teams = [...gameTeams];
+    //     if(teams[0].id === teamId){
+    //         teams[0].home = true;
+    //         teams[1].home = false;
+    //     } else {
+    //         teams[1].home = true;
+    //         teams[0].home = false;
+    //     }
+    //     setGameTeams(teams);
+    // }
 
     const handleGameTypeChange = (e) => {
         setTypeOfGame(e.target.value);
@@ -98,31 +143,31 @@ const EditGameDialog = ({ fights, setFights, game, open, handleClose }) => {
 
     const handleDeleteFight = (e) => {
         
+        let confirmed = window.confirm('Are you sure you want to delete this event? It will be permanently removed from the database')
 
-        //!!!!need to do alert before deleting game
-        deleteFightAPI(e.target.parentElement.id).then(response => {
-            if(response.success){
-                let newFights = fights.filter(elem => {
-                    return elem._id !== e.target.parentElement.id;
-                });
-                setDeleteSuccessful(true);
-                setFights(newFights);
-                setTimeout(() => {
-                    setDeleteSuccessful(false);
-                }, 1500);
-            } else {
-                setDeleteUnsuccessful(true);
-                setTimeout(() => {
-                    setDeleteSuccessful(false);
-                }, 2500);
-            }
-        });
-
-        
+        if(confirmed){
+            deleteFightAPI(e.target.parentElement.id).then(response => {
+                if(response.success){
+                    let newFights = fights.filter(elem => {
+                        return elem._id !== e.target.parentElement.id;
+                    });
+                    setDeleteSuccessful(true);
+                    setFights(newFights);
+                    setTimeout(() => {
+                        setDeleteSuccessful(false);
+                    }, 1500);
+                } else {
+                    setDeleteUnsuccessful(true);
+                    setTimeout(() => {
+                        setDeleteSuccessful(false);
+                    }, 2500);
+                }
+            });
+        }        
     }
 
     return (
-        <Dialog open={open} onClose={handleClose}>
+        <Dialog open={open} onClose={handleClose} disableEnforceFocus>
             <DialogTitle>Edit Game</DialogTitle>
             <DialogContent>
             <DialogContentText >
@@ -147,9 +192,9 @@ const EditGameDialog = ({ fights, setFights, game, open, handleClose }) => {
                 <Grid item sm={5} xs={12}>
                     <TeamSearch team={game.teams[1]} updateTeam={updateTeams} />
                 </Grid>
-                <Grid className={classes.homeTeamRadio} align='center' item xs={12}>
-                    <HomeTeamRadio teams={game.teams} updateHomeTeam={updateHomeTeam} />
-                </Grid>
+                {/* <Grid className={classes.homeTeamRadio} align='center' item xs={12}>
+                    <HomeTeamRadio teams={gameTeams} updateHomeTeam={updateHomeTeam} />
+                </Grid> */}
             
                 <Grid item xs={12}>
                     <TextField 
@@ -193,7 +238,7 @@ const EditGameDialog = ({ fights, setFights, game, open, handleClose }) => {
                                             <Typography >{fight.eventDescription}</Typography>
                                         </Grid>
                                         <Grid item xs={2}>
-                                            <DeleteIcon id={fight._id} onClick={handleDeleteFight} />
+                                            <DeleteIcon className={classes.deleteIcon} id={fight._id} onClick={handleDeleteFight} />
                                         </Grid>
                                         
                                     </Paper>
@@ -208,7 +253,7 @@ const EditGameDialog = ({ fights, setFights, game, open, handleClose }) => {
                                         <Typography>{fight.players[0].lastName} vs. {fight.players[1].lastName}</Typography>    
                                     </Grid>
                                     <Grid item xs={2}>
-                                        <DeleteIcon id={fight._id} onClick={handleDeleteFight} />
+                                        <DeleteIcon className={classes.deleteIcon} id={fight._id} onClick={handleDeleteFight} />
                                     </Grid> 
                                     
                                 </Paper>
@@ -216,10 +261,16 @@ const EditGameDialog = ({ fights, setFights, game, open, handleClose }) => {
                         )
                     })}
                     {deleteSuccessful && 
-                        <Alert severity='success'>Game/Event successfully deleted</Alert>
+                        <Alert severity='success'>Fight/Event successfully deleted</Alert>
                     }
                     {deleteUnsuccessful && 
                         <Alert severity="error">Unable to delete - Try again later</Alert>
+                    }
+                    {updateSuccessful &&
+                        <Alert severity="success">Game has been updated</Alert>                        
+                    }
+                    {updateUnsuccessful &&
+                        <Alert severity="error">Game could not be updated at this time - Try again later</Alert>                        
                     }
             </Grid>
